+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Fix datepicker date format for Custom coverage dates to match current user locale when edit resource. (UIEH-1227)
 * Add tests for useMultiColumnListSort. (UIEH-1183)
 * Fix multiple requests to /resources sent when viewing Titles list on Package Show page. (UIEH-1235)
+* Add tests for PackageShowRoute. (UIEH-1188)
 
 ## [7.0.1] (https://github.com/folio-org/ui-eholdings/tree/v7.0.1) (2021-10-19)